feat auto update sdk

+++ volc-sdk-java/src/main/java/com/volcengine/model/response/ActivationRiskListSampleDataPartResponse.java
@@ -17,6 +17,6 @@ public class ActivationRiskListSampleDataPartResponse {
         @JSONField(name = "PartIds")
         private List<Integer> partIds;
         @JSONField(name = "TotalNum")
-        private Integer totalNum;
+        private Long totalNum;
     }
 }
